feat : player 테이블 삭제
