+++ src/main/resources/data.sql
@@ -1,8 +1,3 @@
-CREATE TABLE player
-(
-    name varchar(30) PRIMARY KEY
-);
-
 CREATE TABLE game
 (
     id          int PRIMARY KEY AUTO_INCREMENT,
@@ -12,11 +7,10 @@ CREATE TABLE game
 
 CREATE TABLE record
 (
+    player_name varchar(30),
     game_id     int,
     position    int     NOT NULL,
     is_winner   boolean NOT NULL,
-    player_name varchar(30),
-    PRIMARY KEY (game_id, player_name),
-    FOREIGN KEY (game_id) REFERENCES game (id),
-    FOREIGN KEY (player_name) REFERENCES player (name)
+    PRIMARY KEY (player_name, game_id),
+    FOREIGN KEY (game_id) REFERENCES game (id)
 );
