+++ src/test/java/racingcar/repository/RecordDaoTest.java
@@ -29,7 +29,6 @@ public class RecordDaoTest {
 
         jdbcTemplate.execute("DROP TABLE record IF EXISTS");
         jdbcTemplate.execute("DROP TABLE game IF EXISTS");
-        jdbcTemplate.execute("DROP TABLE player IF EXISTS");
 
         jdbcTemplate.execute("CREATE TABLE game (\n" +
                 "    id int PRIMARY KEY AUTO_INCREMENT,\n" +
@@ -37,28 +36,19 @@ public class RecordDaoTest {
                 "    game_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP\n" +
                 ");");
 
-        jdbcTemplate.execute("DROP TABLE player IF EXISTS");
-        jdbcTemplate.execute("CREATE TABLE player (\n" +
-                "    name varchar(30) PRIMARY KEY\n" +
-                ");");
-
         jdbcTemplate.execute("DROP TABLE record IF EXISTS");
-        jdbcTemplate.execute("CREATE TABLE record (\n" +
-                "    game_id int,\n" +
-                "    position int NOT NULL,\n" +
-                "    is_winner boolean NOT NULL,\n" +
+        jdbcTemplate.execute("CREATE TABLE record\n" +
+                "(\n" +
                 "    player_name varchar(30),\n" +
-                "    PRIMARY KEY (game_id, player_name),\n" +
-                "    FOREIGN KEY (game_id) REFERENCES game (id),\n" +
-                "    FOREIGN KEY (player_name) REFERENCES player (name)\n" +
+                "    game_id     int,\n" +
+                "    position    int     NOT NULL,\n" +
+                "    is_winner   boolean NOT NULL,\n" +
+                "    PRIMARY KEY (player_name, game_id),\n" +
+                "    FOREIGN KEY (game_id) REFERENCES game (id)\n" +
                 ");");
 
         List<Object[]> trial_count = Arrays.asList(new String[]{"10"}, new String[]{"20"});
         jdbcTemplate.batchUpdate("INSERT INTO game(trial_count) VALUES (?)", trial_count);
-
-        List<Object[]> names = Stream.of(new String[]{"doggy"}, new String[]{"power"})
-                .collect(Collectors.toList());
-        jdbcTemplate.batchUpdate("INSERT INTO player(name) VALUES (?)", names);
 
         List<Object[]> records = Stream.of(new Object[]{1, 8, false, "doggy"}, new Object[]{2, 7, true, "power"})
                 .collect(Collectors.toList());
@@ -83,14 +73,6 @@ public class RecordDaoTest {
         assertThatThrownBy(
                 () -> recordDao.insert(3, false, doggy)
         );
-    }
-
-    @Test
-    @DisplayName("존재하지 않는 player_name이 들어온 경우 예외가 발생한다")
-    void 존재하지_않는_player_name이_들어온_경우_예외가_발생한다() {
-        Vehicle newPlayer = new Vehicle("doggy1", 8);
-
-        assertThatThrownBy(() -> recordDao.insert(2, false, newPlayer));
     }
 
     @Test
